enable float support

+++ custom_helper_op/cc/kernels/sparse_conv3d_kernels_gpu.cu.cc
@@ -215,8 +215,8 @@ void SparseConv3DFunctor<Eigen::GpuDevice, T, SPARSE_CONV3D_FIX_PARAMETOR_ARG_LI
                       SPARSE_CONV3D_DEFINE_INSTANCE_WITH_DILATIONS(instance_type, 1, 1, 1) \
                       SPARSE_CONV3D_DEFINE_INSTANCE_WITH_DILATIONS(instance_type, 2, 2, 2)
 
-// SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(float);
-SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(double); 
+SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(float);
+// SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(double); 
 
 #undef SPARSE_CONV3D_DEFINE_INSTANCE
 
@@ -475,8 +475,8 @@ void SparseConv3DGradFunctor<Eigen::GpuDevice, T, dynamic_default, SPARSE_CONV3D
                 template struct SparseConv3DGradFunctor<GPUDevice, instance_type, true, f_h, f_w, f_d, d_h, d_w, d_d, -1, -1, -1>; \
                 template struct SparseConv3DGradFunctor<GPUDevice, instance_type, false, f_h, f_w, f_d, d_h, d_w, d_d, -1, -1, -1>;
 
-// SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(float);
-SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(double); 
+SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(float);
+// SPARSE_CONV3D_DEFINE_INSTANCE_WITH_TYPE(double); 
 
 #undef SPARSE_CONV3D_DEFINE_INSTANCE
